+++ lib/src/favorite/favorite_screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_github_app/widgets/list_favorite_repos.dart';
+import 'package:flutter_github_app/src/favorite/list_favorite_repos.dart';
 
 class FavoriteScreen extends StatelessWidget {
   @override
