+++ lib/src/home/widgets/search_bar.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_github_app/providers/list_repos_provider.dart';
+import 'package:flutter_github_app/src/home/providers/list_repos_provider.dart';
 import 'package:provider/provider.dart';
 
 class SearchBar extends StatefulWidget {
